Update 791A - Bear and Big Brother.cpp

+++ A-Problems/791A - Bear and Big Brother.cpp
@@ -3,11 +3,15 @@
 using namespace std;
  
 int main() {
-    short a,b,k=0;
+    int a, b, k = 0;
     cin >> a >> b;
-    while (a<=b) {
-        a*=3;b*=2;k++;
+ 
+    while (a <= b) {
+        a *= 3;
+        b *= 2;
+        k++;
     }
+ 
     cout << k;
     return 0;
 }
